admin page UI corrected
